Menu Transition Working
The menu transition are now working, as I see, the button had to be assigned on the inspector because otherwise it didn't work.

Is as simple as add the Transition prefab to the scene and create the load method in the SceneTransitionHandler.

I didn't set this to the NewGame button because I dont't want to make any change on the InGame scene.